Changed controllers to a separate folder
