+++ University_Simulator/src/main/java/com/challenge/alkemy/controllers/AdminController.java
@@ -1,4 +1,4 @@
-package com.challenge.alkemy;
+package com.challenge.alkemy.controllers;
 
 import org.springframework.stereotype.Controller;
