+++ University_Simulator/src/main/java/com/challenge/alkemy/controllers/LoginController.java
@@ -1,4 +1,4 @@
-package com.challenge.alkemy;
+package com.challenge.alkemy.controllers;
 
 import org.springframework.security.authentication.AnonymousAuthenticationToken;
 import org.springframework.security.core.Authentication;
