+++ University_Simulator/src/main/java/com/challenge/alkemy/controllers/StudentController.java
@@ -1,4 +1,4 @@
-package com.challenge.alkemy;
+package com.challenge.alkemy.controllers;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
